changed NavLink to NavLinkCustom, moved pagi-data to pizzas-data and removed pagi, created navlinks.config and removed navlinks-data from util removed addresses-data from util
