+++ src/components/pizza/PizzasList.tsx
@@ -1,5 +1,5 @@
 import { useAppSelector } from 'hooks/use-redux';
-import { getPizzasToRender } from 'util/pagination-data';
+import { getPizzasToRender } from 'util/pizzas-data';
 import { generateRecipeClassName } from 'util/className-generators';
 import { paginationSelectors } from 'redux/reducers/pagination-reducer';
 import PizzaItem from './PizzaItem';
