+++ src/shared/ui/LoadingSpinner.scss
@@ -11,7 +11,8 @@
         left: 0;
         height: 100vh;
         width: 100%;
-        background: rgba(255, 255, 255, 0.9);
+        // background: rgba(255, 255, 255, 0.9);
+        background: rgba(0, 0, 0, 1);
     }
 }
