--- src/util/addresses-data.ts
@@ -1,4 +1,0 @@
-// export const createAddressData = (inputs) => {
-
-// };
-export {};
